+++ CHANGELOG.md
@@ -6,12 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Add `get_bucket_name_or_none_from_bucket_id` to `AccountInfo` and `Cache`
+
 ## [1.7.0] - 2021-04-22
 
 ### Added
 * Add `__slots__` and `__eq__` to `FileVersionInfo` for memory usage optimization and ease of testing
 * Add support for SSE-C server-side encryption mode
-* Add `get_bucket_name_or_none_from_bucket_id` to `AccountInfo` and `Cache`
 
 ### Changed
 * `BasicSyncEncryptionSettingsProvider` supports different settings sets for reading and writing